+++ src/main/java/com/example/takearest/controller/RestaurantController.java
@@ -6,6 +6,7 @@ import com.example.takearest.repository.RestaurantRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.annotation.Secured;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
@@ -32,9 +33,17 @@ public class RestaurantController {
                 .orElseThrow(() -> new RestaurantNotFoundException(id));
     }
 
+    @GetMapping("search/by-name")
+    public ResponseEntity<Restaurant> findByName(@RequestParam String name) {
+        return restaurantRepository.findByName(name)
+                .map(restaurant -> new ResponseEntity<>(restaurant, HttpStatus.OK))
+                .orElseThrow(() -> new RestaurantNotFoundException(name));
+    }
+
     @PostMapping
+    @Secured("ROLE_ADMIN")
     @Transactional
-    public ResponseEntity<Restaurant> save(@RequestBody Restaurant restaurant) {
+    public ResponseEntity<?> save(@RequestBody Restaurant restaurant) {
         Restaurant newRestaurant = restaurantRepository.save(restaurant);
 
         URI uriOfNewResource = ServletUriComponentsBuilder.fromCurrentContextPath()
@@ -48,6 +57,7 @@ public class RestaurantController {
     }
 
     @PutMapping("{id}")
+    @Secured("ROLE_ADMIN")
     @Transactional
     public ResponseEntity<Restaurant> update(@RequestBody Restaurant newRestaurant, @PathVariable long id) {
         Restaurant updatedRestaurant = restaurantRepository.findById(id)
@@ -55,10 +65,7 @@ public class RestaurantController {
                     restaurant.setName(newRestaurant.getName());
                     return restaurantRepository.save(restaurant);
                 })
-                .orElseGet(() -> {
-                    newRestaurant.setId(id);
-                    return restaurantRepository.save(newRestaurant);
-                });
+                .orElseGet(() -> restaurantRepository.save(newRestaurant));
 
         URI uriOfNewResource = ServletUriComponentsBuilder.fromCurrentContextPath()
                 .path(REST_URL + "/{id}")
@@ -70,7 +77,11 @@ public class RestaurantController {
     }
 
     @DeleteMapping("{id}")
+    @Secured("ROLE_ADMIN")
     public ResponseEntity<?> delete(@PathVariable long id) {
+        if (!restaurantRepository.findById(id).isPresent()) {
+            throw new RestaurantNotFoundException(id);
+        }
         restaurantRepository.deleteById(id);
         return ResponseEntity.noContent().build();
     }
